requirements.txt
# Django Core
Django==5.2.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0

# CORS
django-cors-headers==4.5.0

# Filtros y búsqueda
django-filter==24.3

# Documentación API
drf-yasg==1.21.8

# WebSockets
channels==4.1.0
channels-redis==4.2.0

# Base de datos y utilidades
psycopg2-binary==2.9.9  # Para PostgreSQL (opcional)

# Desarrollo y debugging
django-extensions==3.2.3

# Archivos de imagen (opcional)
Pillow==10.4.0

# Variables de entorno
python-decodejwt==0.1.0

# Redis (para WebSockets)
redis==5.1.1